+++ controllers/productController.js
@@ -1,13 +1,13 @@
-const { Products } = require("../models");
-const Error = require("../helpers/error");
-const Response = require("../helpers/response");
+const { Products } = require('../models');
+const Error = require('../helpers/error');
+const Response = require('../helpers/response');
 
 class ProductController {
   async get(req, res, next) {
     try {
       const dataProduct = await Products.findAll({});
       if (dataProduct.length < 1) {
-        throw new Error(400, "There is no product yet");
+        throw new Error(400, 'There is no product yet');
       }
       return new Response(res, 200, dataProduct);
     } catch (error) {
@@ -18,6 +18,7 @@ class ProductController {
     try {
       const { productName, quantity, type } = req.body;
       const createProduct = await Products.create({
+        userID: req.user.id,
         productName,
         quantity,
         type,
@@ -35,8 +36,11 @@ class ProductController {
         where: { id: id },
       });
       if (!searchID) {
-        throw new Error(400, `There is no product with ID ${id}`);
-      }
+        throw new Error(400, `There is no product with ID ${id}`)
+      };
+      if (searchID.userID !== req.user.id) {
+        throw new Error(401, 'Unauthorized to make changes')
+      };
       const updateUser = await Products.update(
         {
           productName: productName,
@@ -55,19 +59,19 @@ class ProductController {
   }
   async deleteByID(req, res, next) {
     try {
-      // const token = req.header('auth-token');
-      // const decode = jwtDecode(token);
-      const { id } = req.params;
-      const dataProduct = await Products.findAll({
+      const dataProduct = await Products.findOne({
         where: { id: id },
       });
       if (!dataProduct) {
         throw new Error(400, `There is no product with ID ${id}`);
       }
+      if (dataProduct.userID !== req.user.id) {
+        throw new Error(401, 'Unauthorized to make changes')
+      };
       const deleteProduct = await Products.destroy({
         where: { id: id },
       });
-      return new Response(res, 200, "Product has been deleted");
+      return new Response(res, 200, 'Product has been deleted');
     } catch (error) {
       next(error);
     }
